- Fixed all unavoidable warnings. - Fixed a few bugs: fdbuf issue and the initialization of 'cstring' in ol_lexer.lex - Added new pel command 'popall', which takes what remains on the stack and pops it to the result tuple - Added marshaling and unmarshaling of opaque value types... Simply copies the previous fdbuf.data using 'xdr_opaque'. - Rewrote elements for fragmentation and defragmentation of tuples. This must have been one of my first elements... It was horribly wrong. With these elements now working I am able to transfer the entire chord.plg program (about 6 fragments of MTU bytes) to a node, where it is of course compiled into an edit and installed into the running instance. - Finalized the p2stub python script that will start a skeleton P2 node with overlog and dataflow installer elements plugged into the 0 port of the dynamic demux and dynamic round robin. - Other minor things that just needed to be done but are not worth mentioning.

+++ trunk/elements/mux.C
@@ -25,7 +25,7 @@ Mux::Mux(string name,
     _timeCallback(NULL),
     _callback(boost::bind(&Mux::callback, this))
 {
-  for (int i = 0;
+  for (unsigned i = 0;
        i < ninputs();
        i++) {
     _pushCallbacks.push_back(0);
@@ -48,7 +48,7 @@ void Mux::catchUp()
   _timeCallback = NULL;
 
   // Go through all pending tuples
-  for (int i = 0; i < ninputs(); i++) {
+  for (unsigned i = 0; i < ninputs(); i++) {
     if (_inputTuples[i] != NULL) {
       // Found one.  Push it out
       assert(_pushCallbacks[i]);
@@ -67,7 +67,7 @@ void Mux::catchUp()
 
   // Wunderbar, we have gotten rid of all of our pending tuples.  Go
   // through and wake up everybody
-  for (int i = 0;
+  for (unsigned i = 0;
        i < ninputs();
        i++) {
     if (_pushCallbacks[i]) {
@@ -84,7 +84,7 @@ void Mux::catchUp()
 
 int Mux::push(int port, TuplePtr p, b_cbv cb)
 {
-  assert((port >= 0) && (port < ninputs()));
+  assert((port >= 0) && (unsigned(port) < ninputs()));
 
   // Is my output blocked?
   if (_blocked) {
